backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
openai
pymupdf
python-docx
python-dotenv
